refactor(panel-api): 패널 목록 가져오기 API 타입 정의한 위치 옮긴다
